Remove ESLint and plugins from npm globals
I prefer Prettier, but sometimes have to use ESLint for team projects.

ESLint works best when installed per-project via package.json. Even if
ESLint is installed globally, plugins must be installed locally, as the
docs explain:

> It is also possible to install ESLint globally rather than locally
> (using npm install eslint --global). However, this is not recommended,
> and any plugins or shareable configs that you use must be installed
> locally in either case.

If ESLint and plugins are installed in the project directory, the
following settings can be added to the VSCode settings.json, for use
with the VSCode ESLint extension and eslint-plugin-vue.

```jsonc
// VSCode settings.json
{
  "eslint.autoFixOnSave": false,
  "eslint.enable": true,
  "eslint.validate": ["javascript", "javascriptreact", "vue"]
}
```

I typically use Prettier instead of ESLint to auto-fix files.

I do still retain a global _.eslintrc_ in my dotfiles repo, because it
is a useful template for sensible defaults.