Add Karabiner keymap for Dygma Raise
These mods enable me to have arrow keys in the default layer (layer 0).

The arrow key locations are different from the default in layer 1.

The right arrow key is currently absent. I plan to use the LEDEFF.NEXT
(LED effects) key for this, but I can't remap it with Karabiner because
it's used by the keyboard itself and not passed through to the OS at
this time. I will eventually be able to use the Bazecor keymapping
software for this. https://github.com/Dygmalab/Bazecor

There is no tilde/backtick key, so I remapped escape.