Restore VSCodium terminal block cursor
I was having an issue where the top and bottom borders of the block were
not showing up when focus moved off the terminal. Not sure if it was
the application itself or my zoom settings. It seems to be better now.

This commit will also restore auto-selection of the terminal renderer.