Dock VSCodium debugger toolbar
I much prefer docked in the sidebar over floating.

I would also prefer to turn off the editor toolbar (the set of icons
docked to the right of the editor tabs) but I don't think there's a
setting for that right now. The "Run Python file in terminal" icon is
the same as the "Start debugging" icon (a play icon). I frequently press
the wrong icon. As a stopgap, I can turn off the GitLens editor toolbar
icons to reduce the size of the toolbar.