Add bootstrap script
https://github.com/MikeMcQuaid/dotfiles
https://github.com/MikeMcQuaid/strap
https://mikemcquaid.com/2016/06/15/replacing-boxen/

I have been using strap to bootstrap my macOS development environment.
The strap script has been very useful, but it is macOS-specific, and I
would like to be able to set up a Linux development environment in a
similar way.

This commit will add bootstrap.sh, which is an initial attempt at
updating strap to work on both macOS and Linux.

There is also some preliminary logic for working with GitHub Codespaces.