Check VSCodium extension versions on Open VSX
VSCodium/vscodium#483
VSCodium/vscodium#494
https://github.com/eclipse/openvsx/wiki/Registry-API

The VSCodium update associated with VSCodium/vscodium#483 required
reinstallation of extensions from the Open VSX registry, but there was
no notification provided to prompt reinstallation, so extensions had
gotten out-of-date.

Open VSX has a nice API with Swagger/OpenAPI docs, and the API can be
queried to check extension metadata. This commit will update the
VSCode/VSCodium extension install script to check Open VSX and ensure
the latest extension versions are installed.

Note that it doesn't appear to be currently possible to install specific
extension versions with the VSCodium CLI and Open VSX. Attempting to run
`code --install-extension "$EXTENSION@$VERSION"` for a given extension
and version returns a `500` error.