Switch VSCode extension script from jq to fx
https://github.com/nodejs/node/releases/tag/v15.3.0
https://github.com/antonmedv/fx
https://github.com/stedolan/jq

jq hasn't had a release in two years, and I'm not yet able to build it
from source on Apple Silicon. fx is easier to use than jq, and depends
on Node.js, which can build from source on Apple Silicon as of v15.3.0.
trentm/json is also a good option.