Install from requirements.txt with pipx
pipx is a helpful tool for managing user Python packages. One limitation
is that pipx doesn't support package installation from requirements.txt.
There's been some discussion of this in pipxproject/pipx#359.

This commit will add a simple Bash script that reads requirements.txt
and installs the packages therein with pipx. If the packages are already
on `$PATH` (such as from installation with Homebrew), they are skipped.
If the packages were installed with pipx on an older version of Python,
they will be reinstalled. I sometimes need to run `pipx reinstall-all`
after updating to a new minor version of Python, and this script will
approximate the behavior of `pipx reinstall-all`.

At the moment, this script doesn't support all requirements.txt syntax,
such as version specification. I don't normally need to pin versions of
pipx packages, so this is fine for me.