Move Brewfile to dotfiles repo
This commit will add https://github.com/br3ndonland/homebrew-brewfile
and update it for Apple Silicon. There isn't really any need to have a
separate repo just for a Brewfile. It was set up that way to support
https://github.com/MikeMcQuaid/strap. Now that a custom bootstrap.sh is
being maintained here, and used in place of strap, the Brewfile can be
moved into this repo.