Set `POETRY_HOME` before running install-poetry.py
br3ndonland/dotfiles@1d8eee0

Commit 1d8eee0 updated the Poetry configuration for install-poetry.py,
the new install script. This commit provides a minor follow-up update.

When bootstrap.sh and strap-after-setup run for the first time, the
exports in .zshrc may not be loaded yet. Sourcing the .zshrc is not an
ideal solution in cases in which the shell has not yet been set to Zsh.
Instead, the `POETRY_HOME` export will be set before installing Poetry.