Ensure Homebrew Bundle installs run on Linux
This commit will correct the Homebrew Bundle commands in the Linuxbrew
script. Previously, if `"$HOME/.Brewfile"` was not present, the Brewfile
would be downloaded from GitHub and used for installs. However, if
`"$HOME/.Brewfile"` was present, it would merely check, not install.

This commit will update the `brew bundle` command so that installs run
if `"$HOME/.Brewfile"` is present but `brew bundle check` fails.