Add VSCode `git.autoRepositoryDetection` setting
VSCode auto-detects Git repositories based on the editors open in the
workspace. This can often result in addition of unwanted repositories.
For example, if Python is installed with Homebrew, and one of the
Python standard library modules is opened, VSCode will add the Homebrew
repository to the workspace.

This commit will set `"git.autoRepositoryDetection": "subFolders"`.
VSCode will only add Git repositories from subfolders of the workspace.