Correct `check_open_vsx` in `vscode-extensions.sh`
br3ndonland/dotfiles@49193b3
https://github.com/eclipse/openvsx/wiki/Registry-API

The `vscode-extensions.sh` script can be used to install extensions for
various VSCode distributions, including VSCodium. It includes a feature
for checking VSCodium extension versions on Open VSX to ensure that all
extensions are up to date. Open VSX has a nice API with Swagger/OpenAPI
docs, and the API can be queried to check extension metadata.

After checking extension versions, the script was using `$EDITOR` to
identify the command used to install extensions. This was problematic
for two reasons:

1. The `--wait` flag should be added to the editor name when setting
  `$EDITOR`, like `codium --wait`, but the command in the script should
  be the name of the editor itself, like `codium`.
2. `$EDITOR` might not be VSCodium.

Additionally, `check_open_vsx` was incorrectly running when extensions
weren't already installed due to the positioning of an `if` conditional.

This commit will correct the logic used to check Open VSX.