Update Brewfile for nonadmin user installs
Some applications can be installed in user-specific directories instead
of in `/Applications`. If installed and run using nonadmin ("standard")
users, it can help reduce privilege when these apps are running.

This commit will add syntax to Brewfile that checks for user `nonadmin`,
and if the user exists, will install to `/Users/nonadmin/Applications`.

The use of `args` syntax with inline `if` statements is undocumented,
and was figured out with trial and error.

There is also an alternative multi-line if statement syntax. See
Homebrew/homebrew-bundle@d9e5ce5, Homebrew/homebrew-bundle@38024cc.