Restart Homebrew services when formulae change
Some Homebrew formulae contain daemon-style background processes that
Homebrew calls "services." Installed formulae with service capabilities
can be listed with the command `brew services list`. When Homebrew
formulae are changed, such as when installing or updating with commands
like `brew bundle install`, it helps to restart the associated services.

This commit will add `restart_service: :changed` to Brewfile so that
services are restarted when the formulae change. There are only two
formulae currently in Brewfile that have background services: PostgreSQL
and Black.