Remove Backblaze Homebrew Cask from Brewfile
When installing the Brewfile with `require_sha: true`, Homebrew reports
an error when trying to install the Backblaze cask:

```text
Error: Cask 'backblaze' does not have a sha256 checksum defined and was not installed.
This means you have the --require-sha option set, perhaps in your HOMEBREW_CASK_OPTS.
```

This is apparently because Backblaze does not offer versioned download
URLs, so the checksum from the unversioned download URL will change
each time a new version is released (Homebrew/homebrew-cask#110134).

This cask can be annoying to install for other reasons. The installer
has to be run manually by navigating to `$HOMEBREW_PREFIX/Caskroom/`,
so it's not much more effort to just download directly from Backblaze.

For these reasons, the Backblaze cask will be removed from the Brewfile.