Use `$STRAP_DOTFILES_BRANCH` in Brewfile URL
The `run_brew_installs` function in `bootstrap.sh` checks for a local
Brewfile first. If a Brewfile is not found, it will attempt to load a
Brewfile from GitHub. The URL to the Brewfile previously had `HEAD`
hard-coded as the Git ref.

This commit will use `$STRAP_DOTFILES_BRANCH` if set, falling back to
`HEAD` if unset.