Add datetime to macOS defaults backup file name
The `macos.sh` script can be used to set macOS defaults. Before changing
any defaults, the script backs up the current defaults to a file.

This commit will add a datetime to the filename, so that the script can
run multiple times and avoid overwriting the previous file each time.