Note macOS `defaults read` permissions requests
The `macos.sh` script can be used to set macOS defaults. Before changing
any defaults, the script backs up the current defaults to a file.

When reading defaults with `defaults read`, macOS will read defaults for
many apps, including calendar or contacts. The terminal may therefore
request permissions to access calendar or contacts. The permissions
requests can be a bit disconcerting when it's unclear what they are for,
so this commit will add a print command explaining what's going on.