Rename VSCode Microsoft marketplace extension file
VSCode extensions are listed in plain-text files in `vscode/extensions/`
and can be installed with `scripts/vscode-extensions.sh`. There are two
text files, one for extensions that can be installed from the Open VSX
marketplace, and another for extensions not listed in the Open VSX
marketplace that must be installed from the Microsoft marketplace
instead. The Microsoft marketplace itself is proprietary, but extensions
in the marketplace may or may not be.

This commit will update the name of the text file containing Microsoft
marketplace extensions with the owner of the marketplace (Microsoft) to
avoid any ambiguity about whether the extensions in the file are
proprietary or not.