Alias `gsed` to `sed`
macOS includes BSD `sed`, which is different from GNU `sed`. GNU `sed`
is installed with Homebrew, which adds it to `$PATH` as `gsed`.

This commit will create an alias from `gsed` to `sed` to ensure that GNU
`sed` is always used. Aliases need to be set for each shell, so they
will be added to both `.zshrc` and `.bashrc`.