Update nonzero exit code statements in `.zshrc`
Some of the logic in `.zshrc` for setting up Homebrew added in 7b5d074
can result in nonzero exit codes. `[[ -z $HOMEBREW_PREFIX ]]` will exit
`1` if `$HOMEBREW_PREFIX` is set (as it is after installing Homebrew).
Zsh handles these errors gracefully, so it's not a problem for `.zshrc`,
but it is good practice to avoid error codes in config files.

This commit will update the Homebrew setup statements so that they don't
exit with error codes.