Remove `git pull --ff-only` from `.gitconfig`
The `.gitconfig` had two settings that were silently conflicting:

1. `rebase = true`
2. `ff = only`

This basically means `git pull --ff-only --rebase`, which doesn't make
sense, because you can't fast-forward if you have to rebase. For some
reason Git doesn't error with these two settings, but it appears that
`--ff-only` wins out. Pulling with `--ff-only` means that if there are
remote changes and the user runs `git pull`, a merge commit is created.
Pull with rebase means that the remote changes will be pulled without a
merge commit, then the local commits are replayed on top of the pulled
commits.

This commit will remove the `ff = only` setting, and will also add the
`branch.autosetuprebase` setting to ensure this behavior is matched when
setting up new branches.

https://www.git-scm.com/docs/git-config
https://www.git-scm.com/docs/git-pull
https://www.git-scm.com/book/en/v2/Git-Branching-Rebasing