Set `core.commentChar` in `.gitconfig`
Git reads lines starting with `#` or `;` as comments by default.
`#` characters have special meanings in some contexts, such as Markdown
formatting (to indicate headers) and issue tracker references for GitHub
and other platforms (indicate reference to issue or PR), so considering
`#` as a comment is not ideal. On the other hand, `;` characters are not
used to indicate Markdown headers or issue tracker references.

This commit will set `core.commentChar` in `.gitconfig` to `;` so that
lines starting with `#` are not misinterpreted as comments.

Note that VSCode syntax highlighting doesn't respect `core.commentChar`.
microsoft/vscode#37387