Revert "Pin VSCode Python extension to avoid Jedi snippets"
This reverts commit 02033b85d209a611eabe99e2216178cd4c68acdb.

The VSCode Python extension altered the behavior of snippet expansion
when using the open-source Jedi LSP, adding parentheses after class
names. This behavior may be bothersome when adding type annotations.
To avoid this behavior, the VSCode Python extension was pinned to a
prior version, `ms-python.python@2021.9.1246542782`.

Thanks to a community contribution, this issue is now resolved as of
ms-python.python@2023.10.0 (2023-06-08).

microsoft/vscode-python#15858