Add new VSCode Flake8 and mypy extensions
This commit will replace the settings for Flake8, mypy, and Pylint that
were previously managed by the main `ms-python.python` extension with
updated settings for the new extensions `ms-python.flake8` and
`ms-python.mypy-type-checker`.

This commit will also add the new extensions to the appropriate manifest
file. The extensions are available on Open VSX for use with VSCodium.

https://open-vsx.org/extension/ms-python/flake8
https://open-vsx.org/extension/ms-python/mypy-type-checker