Don't start services by default
The Python Black formatter includes a local web server feature.
The Homebrew formula for Black starts `blackd` by default, but this is
not necessary. The VSCode Black formatter extension uses LSP instead of
`blackd`, and there are no other uses of `blackd` on the system.

The Homebrew formula for PostgreSQL starts the server by default, but
this is not always needed. The service can be started on-demand with
`brew services run postgresql@<VERSION>`.

Running services can be stopped with `brew services stop --all`.

https://black.readthedocs.io/en/stable/usage_and_configuration/black_as_a_server.html
https://docs.brew.sh/Manpage#services-subcommand
https://github.com/Homebrew/homebrew-core/blob/3a8a9cb5fa04640233243b0dfa4550a48f0548f2/Formula/b/black.rb
https://github.com/microsoft/vscode-black-formatter
https://github.com/microsoft/vscode-python/issues/3336