Configure kitty `paste_actions`
kitty 0.31.0 (2023-11-08) added a confirmation dialog that appears when
pasting text into the terminal, with options for "sanitize and paste" or
"paste anyway". The `paste_actions` configuration setting can be used to
adjust this behavior.

This commit will update `paste_actions` from the default `configure` to
`confirm-if-large,replace-dangerous-control-codes` to disable the prompt
while still performing some security checks.

The `replace-newline` action will not be used because it often breaks
commands that contain newlines (`\n`). Note that part of the motivation
for the `replace-newline` action was a CRLF injection vulnerability
(CVE-2008-2383), but this vulnerability appears to have been addressed
by kovidgoyal/kitty@60a7a53, so it's not clear if allowing pasting of
newlines represents any additional security risk.

kovidgoyal/kitty@60a7a53
CVE-2008-2383
https://github.com/advisories/GHSA-94vw-2f3v-j88m
https://sw.kovidgoyal.net/kitty/changelog/
https://sw.kovidgoyal.net/kitty/conf/