Enable VSCode Git commit input validation
Note that, when this setting was introduced in VSCode 1.20, the default
was "warn", but apparently the default was later changed to "off".

Also note that the setting type may be changing from string to Boolean
in an upcoming release (microsoft/vscode@9b0d743).

https://code.visualstudio.com/docs/sourcecontrol/overview
https://code.visualstudio.com/updates/v1_20
microsoft/vscode@79f8189
microsoft/vscode@9b0d743