Use split VSCode Git diff editor
VSCode/VSCodium has a Git diff viewer/editor. The editor can either be
inline in the same editor (called the "unified view" on GitHub) or split
into two editors. One of the downsides to the inline editor is that
deleted lines apparently cannot be copied.

This commit will ensure that the split editor is used.