Add alias for first Git commit to `.gitconfig`
This commit will add an alias to the `.gitconfig` that can show the
first commit in the repo. All commits can be shown earliest first with
`git log --reverse`, but this includes more than just the first commit.
The `git log` command doesn't have a built-in feature for this purpose,
but `git rev-list` can be used.

https://www.git-scm.com/docs/git-rev-list