Remove Terraform from Brewfile
The `terraform@1.5` formula was deemed "stale" and removed.

https://github.com/Homebrew/homebrew-core/pull/183818

The `terraform` formula has been deprecated in Homebrew because of
license changes. The Homebrew formula page says,

> We will not accept any new Terraform releases in homebrew/core
> (with the BUSL license).
> The next release changed to a non-open-source license:
> https://www.hashicorp.com/blog/hashicorp-adopts-business-source-license
> See our documentation for acceptable licences:
> https://docs.brew.sh/License-Guidelines

HashiCorp relicensed Terraform from the (Mozilla Public License) to the
non-open-source BSL (Business Source License). Terraform 1.5 was the
final release to be licensed under MPL.

OpenTofu is an open-source alternative to HashiCorp Terraform. It was
initially called OpenTF.

https://formulae.brew.sh/formula/opentofu
https://formulae.brew.sh/formula/terraform
https://github.com/opentofu/opentofu
https://opentofu.org/
https://www.hashicorp.com/blog/hashicorp-adopts-business-source-license
https://blog.gruntwork.io/the-future-of-terraform-must-be-open-ab0b9ba65bca