Revert "Fix prompt character for kitty 0.33.1"
This reverts commit 6939b77082383816982877046755502bda07cd41.

An update in the kitty 0.33.1 release broke macOS symbol font fallback,
resulting in the ❯ prompt character (U+276f) displaying as an empty box.
In the meantime, the `symbol_map` configuration setting was used to set
an explicit fallback.

https://github.com/kovidgoyal/kitty/issues/7249
https://sw.kovidgoyal.net/kitty/conf.html

Subsequent releases fixed font fallback, so the fallback character no
longer needs to be specified in `kitty.conf`.