Fix arg parsing error in `linuxbrew.sh`
The `linuxbrew.sh` Linux setup script installs some `apt` packages.
In addition to the default array of packages in the script, it offers
the ability to install more packages by passing them as args to the
script. There was a syntax error in which, rather than adding args to
the array of packages, the literal string `[@]` was being added to the
array. This commit will fix the syntax error.