Use `type` instead of `command -v` in scripts
This commit will update `bootstrap.sh` and `linuxbrew.sh` to detect
executable programs with `type` instead of with `command -v`.
The `command` executable does not exist on Linux GitHub Actions runners.
This is strange because `command` is included in the POSIX spec.
https://pubs.opengroup.org/onlinepubs/9799919799/utilities/command.html

To avoid errors, the `type` executable will be used instead. `type -P`
can output the path to the executable, but the `-P` option is specific
to Bash and is not a POSIX default. For portability, `type` will be used
without options where possible.
https://pubs.opengroup.org/onlinepubs/9799919799/utilities/type.html