Add IP checker function to `$fpath`
It is sometimes helpful to check a machine's public IP address from the
command-line. The free ip.me service from Proton is one way to do this.
Running `curl ip.me` will return just the IP address in plain-text, but
other HTTP clients like wget or HTTPie will return the full webpage.
This is due to how ip.me handles the user agent. Although it is not
documented on the curl manpage, curl sets `User-Agent:curl/[version]`,
such as `User-Agent:curl/8.11.0` (the current version). Similar
behavior can therefore be achieved by setting the user agent like curl.

These commands all produce equivalent results:

```sh
curl ip.me
wget -qO - -U curl/8.11.0 ip.me
http ip.me User-Agent:curl/8.11.0 -b # https://httpie.io/
```

This commit will add a script for checking the IP address into a new
directory `.zfunc`. This directory is set up so the function can be
loaded automatically by shells like Bash and Zsh. For Bash, the
directory will simply be added to `$PATH`. For Zsh, the directory will
be added to `$fpath` (`fpath+=($HOME/.zfunc)`), a special array variable
telling Zsh where to look for shell functions, and then automatically
loaded into each shell (`autoload -Uz $HOME/.zfunc/*(:tX)`). `(:tX)` is
a special Zsh expansion parameter that lists the filenames `(:t)` of all
executable files (`X`) in the given directory.

As this IP script may be used by multiple shells, the shebang will be
omitted and a directive will ShellCheck to treat it as POSIX sh.

As part of this addition to the `$fpath`, this commit will also fix a
long-standing error in the `$fpath` introduced in commit 909c1e7 that
reflected a misunderstanding of the `$fpath` and of the difference
between `$fpath` and `$FPATH`. The syntax introduced in 909c1e7 modified
`$fpath`, but with `$FPATH` syntax, which is similar to `$PATH` syntax
(a string with colon-separated values). However, `$fpath` is actually an
array. This commit will correct the syntax error.

VSCode settings will be updated so that files in `.zfunc` are treated as
shell scripts (using the `shellscript` language identifier).

https://everything.curl.dev/http/modify/user-agent.html
https://ip.me/

br3ndonland/dotfiles@909c1e7
https://code.visualstudio.com/docs/languages/identifiers
https://zsh.sourceforge.io/Doc/Release/Expansion.html
https://zsh.sourceforge.io/Doc/Release/Functions.html