Fix dotenv file associations
Commit f317401 updated dotenv file associations to include files that
begin with `env` instead of `.env`. Dotenv files are sometimes named
`environment.env` or simply `env`, so this was somewhat helpful. However
it also meant that files like `environment.md` would be incorrectly
associated with `dotenv` when they should be associated with `markdown`.
This commit will adjust the syntax to use more specific glob patterns,
with curly braces used to group conditions and reduce code duplication.

https://code.visualstudio.com/docs/editor/glob-patterns
https://code.visualstudio.com/docs/languages/identifiers