Use Zsh array join for `$PATH` in `.zshrc`
The line on which the `$PATH` variable is set in `.zshrc` is getting
relatively long (178 characters), and therefore difficult to read.

This commit will transform `$PATH` from `PATH=foo:bar:baz` syntax into
an array, and will then use Zsh parameter expansion to join the array
items to reconstitute `PATH=foo:bar:baz` syntax for export to the shell.

https://zsh.sourceforge.io/Doc/Release/Expansion.html#Parameter-Expansion