Align syntax for starship and mise activation
This commit will update the syntax used in `.bashrc` and `.zshrc` used
to activate the Starship prompt and the mise-en-place tool environment
manager. Either `eval` or `source` can be used to activate these tools.
This commit will align the syntax to use `source` consistently, and will
add a missing mise activation command to `.bashrc`.

https://mise.jdx.dev/
https://starship.rs