Add VSCode snippet for Prettier ignore comments
This commit will add a global snippet and keybinding for inserting
language-specific Prettier ignore comments. It is helpful to make the
snippet global due to the large number of languages Prettier formats.
The `scope` (languages to which the snippet applies) match the language
IDs in the `settings.json` that are specified for use with Prettier.

https://code.visualstudio.com/docs/editor/userdefinedsnippets
https://prettier.io/docs/ignore