Switch to new OpenTofu VSCode extension
HashiCorp relicensed Terraform from the (Mozilla Public License) to the
non-open-source BSL (Business Source License). Terraform 1.5 was the
final release to be licensed under MPL.

OpenTofu is an open-source alternative to HashiCorp Terraform. It was
initially called OpenTF.

https://formulae.brew.sh/formula/opentofu
https://github.com/opentofu/opentofu
https://opentofu.org/faq/
https://www.hashicorp.com/blog/hashicorp-adopts-business-source-license
https://blog.gruntwork.io/the-future-of-terraform-must-be-open-ab0b9ba65bca

This commit will remove the HashiCorp Terraform and HCL extensions and
switch to the new official OpenTofu extension for VSCode and VSCodium.
https://github.com/opentofu/vscode-opentofu
https://marketplace.visualstudio.com/items?itemName=OpenTofu.vscode-opentofu
https://open-vsx.org/extension/OpenTofu/vscode-opentofu
https://opentofu.org/blog/opentofu-1-10-0/

The `files.associations` for OpenTofu files (`"**/*.tofu": "terraform"`)
(db877e7) should no longer be necessary because the extension adds
support for `.tofu` files.
https://github.com/opentofu/vscode-opentofu/blob/e50f9640c628b46fbaaad7b0ba769aceca800fdc/package.json#L56-L71