Switch `foxundermoon.shell-format` to `mkhl.shfmt`
The `foxundermoon.shell-format` VSCode extension can be used to format
shell scripts with `shfmt` (https://github.com/mvdan/sh).

The current version of the extension (7.2.8) is unusable because it is
missing a dependency (foxundermoon/vs-shell-format#396). Unfortunately,
it's not practical to roll back to a previous version of the extension
because, prior to 7.2.8, it wasn't published to Open VSX for four years
(foxundermoon/vs-shell-format#394). Note that `mkhl.shfmt` is mentioned
in the `shfmt` README, but `foxundermoon.shell-format` is not.

This commit will switch `foxundermoon.shell-format` to `mkhl.shfmt`.
The `"shfmt.formatIgnored": false` setting is needed because of an issue
with the default (https://codeberg.org/mkhl/vscode-shfmt/pulls/21). In
order to properly ignore files, a `.editorconfig` will be added. `shfmt`
and `mkhl.shfmt` will recognize this file and ignore Zsh scripts.
https://github.com/mvdan/sh/blob/HEAD/cmd/shfmt/shfmt.1.scd

https://github.com/mvdan/sh
https://open-vsx.org/extension/foxundermoon/shell-format
https://open-vsx.org/extension/mkhl/shfmt
https://codeberg.org/mkhl/vscode-shfmt