Add `.gitconfig` alias for pruning local branches
This commit will add a Git alias for pruning local branches. When GitHub
PRs are squash merged, the branch on the remote (on GitHub) is usually
deleted, but the branch on each user's local machine is still there. It
can therefore be tedious to repeatedly delete all the local branches.

The alias added in this commit will parse output of `git branch -vv` to
identify branches for which the remotes are listed as "gone" and then
pass these remoteless branch names into the `git branch -D` command.

```sh
git branch -vv | grep ': gone]' | awk '{print $1}' | xargs -r git branch -D
```