Update Starship config JSON schema
This commit will update the Starship config JSON schema to commit
starship/starship@d85ee0a. This schema is used to lint and format the
`starship.toml` config file with Tombi.

A specific commit is referenced because the hosted schema in the docs
(https://starship.rs/config-schema.json) is often out of date. The
Starship docs appear to be deployed manually with `workflow_dispatch`
only after releases.