Avoid running Homebrew activation code twice
This commit provides a small follow-up to 9f3818b, in which the `$-`
special parameter was used in a conditional to check for interactive
shell sessions (`i`) and only run `mise activate zsh --shims` when the
session is non-interactive.

There is also some Homebrew activation code in `.zprofile`. It seems
harmless to run it twice (once in `.zprofile`, once in `.zshrc`) but it
is also unnecessary.

This commit will wrap the Homebrew activation code in the conditional as
well (`if [[ $- != *i* ]]`).