Simplify `$-` checks for non-interactive shells
This commit will provide a small follow-up to 9f3818b and 5bb247a, in
which the `$-` special parameter was used in a conditional to check for
interactive shell sessions (`i`) and only activate mise shims when the
session is non-interactive.

This commit will simplify the conditional to return if the session is
interactive (`if [[ $- != *i* ]]; then` -> `[[ $- == *i* ]] && return`).