Add `$-` checks for interactive shells
This commit will provide a follow-up to 9f3818b, 5bb247a, and d6cb054,
in which the `$-` special parameter was used in a conditional to check
for interactive shell sessions (`i`) and only activate mise shims when
the session is non-interactive.

This commit will address the converse case for interactive sessions. The
`.bashrc` and `.zshrc` configuration files will be updated so they only
load for interactive shells (`[[ $- != *i* ]] && return`).