Update and unify Zsh and Bash keybindings
This commit will update the keybindings used in the Zsh and Bash shells.

Zsh keybindings will be updated to correspond with zsh4humans v5, with
some modifications to make the keyboard shortcuts more VSCode-like. For
example, Command-Left will be mapped to 'beginning-of-line' and
Command-Right will be mapped to 'end-of-line' rather than to previous
and next word as zsh4humans does.

Corresponding keybindings for Bash will be added to a Readline init file
`.inputrc`. Zsh does not read this file, and Zsh has additional bindings
available in its line editor (`zle`) that are not supported by Readline.
Emacs mode (`set -o emacs`) is typically active by default in Bash, but
"warning: line editing not enabled" may be seen in some cases, so Emacs
mode will be explicitly set in `.bashrc` to ensure consistent behavior.

Note that, while Command-Delete (⌘ ⌦) will be mapped to 'kill-line',
Command-Backspace (⌘ ⌫) will not be mapped to 'backward-kill-line'
because different terminal emulators handle it in different ways. For
example, the VSCode integrated terminal supports Command-Backspace for
'backward-kill-line' out of the box, but kitty does not. Configuration
for specific terminal emulators will be added in subsequent commits.

- https://www.gnu.org/software/bash/manual/html_node/Command-Line-Editing.html
- https://www.gnu.org/software/bash/manual/html_node/Readline-Init-File.html
- https://zsh.sourceforge.io/Guide/zshguide04.html
- https://zsh.sourceforge.io/Doc/Release/Zsh-Line-Editor.html
- https://github.com/romkatv/zsh4humans/blob/v5/z4h.zsh