Map Cmd-Backspace to backward-kill-line in kitty
While Command-Delete (⌘ ⌦) has been successfully mapped to 'kill-line',
Command-Backspace (⌘ ⌫) is not currently mapped to 'backward-kill-line'
because different terminal emulators handle it in different ways. For
example, the VSCode integrated terminal supports Command-Backspace for
'backward-kill-line' out of the box, but kitty does not. kitty offers
the `show-key` kitten to detect and troubleshoot key codes. When running
`kitten show-key -m kitty`, Cmd-Backspace (⌘ ⌫) shows as 'CSI 127 ; 9 u'
('^[[127;9u'), but mapping this sequence to 'backward-kill-line' does
not work. Instead, this commit will update `kitty.conf` to map
Cmd-Backspace (`cmd+backspace` or `super+backspace`) to a text sequence
('\x15'), and apparently, this will automatically bind Command-Backspace
to 'backward-kill-line'. Although there are no changes needed to shell
configuration files, comments will be added to `.inputrc` and `.zshrc`
to explain this behavior.

- https://github.com/kovidgoyal/kitty/issues/264
- https://sw.kovidgoyal.net/kitty/faq/
- https://sw.kovidgoyal.net/kitty/keyboard-protocol/
- https://www.gnu.org/software/bash/manual/html_node/Command-Line-Editing.html
- https://www.gnu.org/software/bash/manual/html_node/Readline-Init-File.html
- https://zsh.sourceforge.io/Doc/Release/Zsh-Line-Editor.html
- https://zsh.sourceforge.io/Guide/zshguide04.html