Update Starship config JSON schema
This commit will update the Starship config JSON schema to tag
starship/starship@v1.24.0. This schema is used to lint and format the
`starship.toml` config file with Tombi.

A specific commit was referenced previously because the hosted schema in
the docs (https://starship.rs/config-schema.json) is often out of date.
The Starship docs appear to be deployed manually with
`workflow_dispatch` only after releases.