Add ZSA Keymapp to Brewfile
This commit will add the Keymapp cask to the Brewfile. Keymapp is the
desktop companion to the ZSA Oryx configurator. It is used to flash
layouts onto keyboards like the ZSA Voyager and can also display keys
being typed in real time.

The app will be a user-specific install into `~/Applications` and will
not require a SHA check because they unfortunately do not provide SHAs.

- https://formulae.brew.sh/cask/keymapp
- https://github.com/Homebrew/homebrew-cask/blob/main/Casks/k/keymapp.rb
- https://www.zsa.io/flash
- https://www.zsa.io/oryx
- https://www.zsa.io/voyager