Update to Hatch 1.16
Hatch 1.16 introduces support for PEP 735 dependency groups in
`pyproject.toml`. Dependency groups allow development dependencies to be
moved out of the `[project.optional-dependencies]` table and into a
separate `[dependency-groups]` table. This is helpful because optional
dependencies are included with package metadata, so previously, groups
of dependencies in the `[project.optional-dependencies]` table (also
called "features" or "extras") were all included in the built package,
and visible in `PKG-INFO` text files in sdists (source distributions),
even if they were only used for development of the project itself. The
`[dependency-groups]` table is not included in built packages, so the
package has cleaner metadata when built and distributed to registries
like PyPI.

https://github.com/pypa/hatch/releases
https://hatch.pypa.io/latest/blog/2025/11/24/hatch-v1160/
https://hatch.pypa.io/latest/config/environment/overview/#dependency-groups
https://peps.python.org/pep-0735/